+++ app/_layout.tsx
@@ -11,6 +11,18 @@ export default function RootLayout() {
       }}
       />
       <Stack.Screen
+      name="splash"
+      options={{
+        headerShown: false
+      }}
+      />
+      <Stack.Screen
+      name="welcome"
+      options={{
+        headerShown: false
+      }}
+      />
+      <Stack.Screen
       name="login"
       options={{
         headerShown: false
