+++ app/login.tsx
@@ -4,7 +4,6 @@ import { View, Text, StyleSheet, Image, Alert, TextInput } from 'react-native';
 import { SafeAreaProvider, SafeAreaView } from 'react-native-safe-area-context';
 import { Button } from 'react-native-paper';
 
-
 export default function Login() {
     const router = useRouter();
