+++ app/splash.tsx
@@ -1,0 +1,28 @@
+import React from "react";
+import { Image, StyleSheet, View } from "react-native";
+
+const Logo = require("../assets/images/logo.png");
+
+export default function Splashscreen() {
+  return (
+    <View style={styles.container}>
+      <View>
+        <Image source={Logo} style={styles.image} />
+      </View>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+    backgroundColor: "#ffffff",
+  },
+  image: {
+    width: 100,
+    height: 100,
+    resizeMode: "cover",
+  }
+})
